Separate LinkServerDatabase into distinct classes
