+++ bot/src/test/kotlin/org/epilink/bot/SessionChecksTest.kt
@@ -18,7 +18,7 @@ import io.ktor.sessions.sessions
 import io.ktor.util.pipeline.PipelineContext
 import io.mockk.*
 import kotlinx.coroutines.runBlocking
-import org.epilink.bot.db.LinkServerDatabase
+import org.epilink.bot.db.LinkDatabaseFacade
 import org.epilink.bot.db.UsesTrueIdentity
 import org.epilink.bot.http.ApiErrorResponse
 import org.epilink.bot.http.LinkSessionChecks
@@ -134,7 +134,7 @@ class SessionChecksTest : KoinBaseTest(
             every { context } returns call
             every { finish() } just runs
         }
-        mockHere<LinkServerDatabase> {
+        mockHere<LinkDatabaseFacade> {
             coEvery { getUser("userid") } returns null
         }
         runBlocking {
@@ -157,7 +157,7 @@ class SessionChecksTest : KoinBaseTest(
         val context = mockk<PipelineContext<Unit, ApplicationCall>> {
             every { context } returns call
         }
-        mockHere<LinkServerDatabase> {
+        mockHere<LinkDatabaseFacade> {
             coEvery { getUser("userid") } returns mockk()
         }
         runBlocking {
@@ -196,7 +196,7 @@ class SessionChecksTest : KoinBaseTest(
             every { context } returns call
             every { finish() } just runs
         }
-        mockHere<LinkServerDatabase> {
+        mockHere<LinkDatabaseFacade> {
             coEvery { isUserIdentifiable("adminid") } returns false
         }
         declare(named("admins")) { listOf("adminid") }
@@ -220,7 +220,7 @@ class SessionChecksTest : KoinBaseTest(
         val context = mockk<PipelineContext<Unit, ApplicationCall>> {
             every { context } returns call
         }
-        mockHere<LinkServerDatabase> {
+        mockHere<LinkDatabaseFacade> {
             coEvery { isUserIdentifiable("adminid") } returns true
         }
         declare(named("admins")) { listOf("adminid") }
